Present "interesting" minutes in log views, drop the minimap (#114)
