+++ src/main/kotlin/io/github/inductiveautomation/kindling/log/LogPanel.kt
@@ -1,9 +1,9 @@
 package io.github.inductiveautomation.kindling.log
 
-import com.formdev.flatlaf.ui.FlatScrollBarUI
 import io.github.inductiveautomation.kindling.core.Detail.BodyLine
 import io.github.inductiveautomation.kindling.core.DetailsPane
 import io.github.inductiveautomation.kindling.core.Filter
+import io.github.inductiveautomation.kindling.core.FilterPanel
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.Advanced.Debug
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.Advanced.HyperlinkStrategy
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.General.ShowFullLoggerNames
@@ -11,7 +11,6 @@ import io.github.inductiveautomation.kindling.core.Kindling.Preferences.General.
 import io.github.inductiveautomation.kindling.core.LinkHandlingStrategy
 import io.github.inductiveautomation.kindling.core.ToolOpeningException
 import io.github.inductiveautomation.kindling.core.ToolPanel
-import io.github.inductiveautomation.kindling.log.LogViewer.ShowDensity
 import io.github.inductiveautomation.kindling.log.LogViewer.TimeStampFormatter
 import io.github.inductiveautomation.kindling.utils.Action
 import io.github.inductiveautomation.kindling.utils.EDT_SCOPE
@@ -23,6 +22,7 @@ import io.github.inductiveautomation.kindling.utils.ReifiedJXTable
 import io.github.inductiveautomation.kindling.utils.VerticalSplitPane
 import io.github.inductiveautomation.kindling.utils.attachPopupMenu
 import io.github.inductiveautomation.kindling.utils.configureCellRenderer
+import io.github.inductiveautomation.kindling.utils.debounce
 import io.github.inductiveautomation.kindling.utils.isSortedBy
 import io.github.inductiveautomation.kindling.utils.selectedRowIndices
 import io.github.inductiveautomation.kindling.utils.toBodyLine
@@ -32,28 +32,16 @@ import kotlinx.coroutines.launch
 import net.miginfocom.swing.MigLayout
 import org.jdesktop.swingx.JXSearchField
 import org.jdesktop.swingx.table.ColumnControlButton.COLUMN_CONTROL_MARKER
-import java.awt.Dimension
-import java.awt.Graphics
-import java.awt.Graphics2D
-import java.awt.Rectangle
-import java.awt.RenderingHints
-import java.awt.geom.AffineTransform
-import java.time.Duration
-import java.time.Instant
 import java.util.Vector
 import javax.swing.Icon
 import javax.swing.JCheckBox
 import javax.swing.JComboBox
-import javax.swing.JComponent
 import javax.swing.JLabel
 import javax.swing.JPanel
 import javax.swing.JPopupMenu
-import javax.swing.JScrollBar
 import javax.swing.ListSelectionModel
 import javax.swing.SortOrder
-import javax.swing.SwingConstants
-import javax.swing.UIManager
-import kotlin.math.absoluteValue
+import kotlin.time.Duration.Companion.milliseconds
 import io.github.inductiveautomation.kindling.core.Detail as DetailEvent
 
 typealias LogFilter = Filter<LogEvent>
@@ -75,8 +63,6 @@ class LogPanel(
 
     private val totalRows: Int = rawData.size
 
-    private val densityDisplay = GroupingScrollBar()
-
     private val header = Header(totalRows)
 
     private val columnList = if (rawData.first() is SystemLogEvent) {
@@ -92,9 +78,7 @@ class LogPanel(
         }
     }
 
-    private val tableScrollPane = FlatScrollPane(table) {
-        verticalScrollBar = densityDisplay
-    }
+    private val tableScrollPane = FlatScrollPane(table)
 
     private val sidebar = FilterSidebar(
         NamePanel(rawData),
@@ -112,8 +96,7 @@ class LogPanel(
             null
         },
         TimePanel(
-            lowerBound = rawData.first().timestamp,
-            upperBound = rawData.last().timestamp,
+            rawData,
         ),
     )
 
@@ -144,37 +127,42 @@ class LogPanel(
         }
     }
 
-    private fun updateData() {
-        BACKGROUND.launch {
-            val selectedEvents = table.selectedRowIndices().map { row -> table.model[row].hashCode() }
-            val filteredData = if (Debug.currentValue) {
-                // use a less efficient, but more debuggable, filtering sequence
-                filters.fold(rawData) { acc, logFilter ->
-                    acc.filter(logFilter::filter).also {
-                        println("${it.size} left after $logFilter")
-                    }
-                }
-            } else {
-                rawData.filter { event ->
-                    filters.all { filter -> filter.filter(event) }
+    private val dataUpdater = debounce(50.milliseconds, BACKGROUND) {
+        val selectedEvents = table.selectedRowIndices().map { row -> table.model[row].hashCode() }
+        val filteredData = if (Debug.currentValue) {
+            // use a less efficient, but more debuggable, filtering sequence
+            filters.fold(rawData) { acc, logFilter ->
+                acc.filter(logFilter::filter).also {
+                    println("${it.size} left after $logFilter")
                 }
             }
-
-            EDT_SCOPE.launch {
-                table.apply {
-                    model = createModel(filteredData)
-
-                    selectionModel.valueIsAdjusting = true
-                    model.data.forEachIndexed { index, event ->
-                        if (event.hashCode() in selectedEvents) {
-                            val viewIndex = convertRowIndexToView(index)
-                            addRowSelectionInterval(viewIndex, viewIndex)
-                        }
-                    }
-                    selectionModel.valueIsAdjusting = false
-                }
+        } else {
+            rawData.filter { event ->
+                filters.all { filter -> filter.filter(event) }
             }
         }
+
+        EDT_SCOPE.launch {
+            table.apply {
+                model = createModel(filteredData)
+
+                selectionModel.valueIsAdjusting = true
+                model.data.forEachIndexed { index, event ->
+                    if (event.hashCode() in selectedEvents) {
+                        val viewIndex = convertRowIndexToView(index)
+                        addRowSelectionInterval(viewIndex, viewIndex)
+                    }
+                }
+                selectionModel.valueIsAdjusting = false
+            }
+        }
+    }
+
+    private fun updateData() = dataUpdater()
+
+    fun reset() {
+        sidebar.filterPanels.forEach(FilterPanel<LogEvent>::reset)
+        header.search.text = null
     }
 
     @Suppress("UNCHECKED_CAST")
@@ -323,71 +311,7 @@ class LogPanel(
         }
     }
 
-    inner class GroupingScrollBar : JScrollBar() {
-        private val density: Map<Instant, Int>
-        private val rangex: Int
-
-        init {
-            val delta = Duration.between(rawData.first().timestamp, rawData.last().timestamp)
-            val slice = delta.dividedBy((rawData.size.toLong() / 60).coerceAtLeast(1))
-            val insertionPoint = DURATIONS.binarySearch { it.compareTo(slice) }
-            val aggregate = DURATIONS[insertionPoint.absoluteValue - 1]
-
-            toolTipText = aggregate.toString()
-
-            density = rawData.groupingBy {
-                it.timestamp.truncatedTo(DurationUnit(aggregate))
-            }.eachCount()
-            rangex = density.values.maxOf { it }
-        }
-
-        override fun getUnitIncrement(direction: Int): Int {
-            return table.getScrollableUnitIncrement(tableScrollPane.viewport.viewRect, SwingConstants.VERTICAL, direction)
-        }
-
-        override fun getBlockIncrement(direction: Int): Int {
-            return table.getScrollableBlockIncrement(tableScrollPane.viewport.viewRect, SwingConstants.VERTICAL, direction)
-        }
-
-        private val customUI = object : FlatScrollBarUI() {
-            override fun paintTrack(g: Graphics, c: JComponent, trackBounds: Rectangle) {
-                super.paintTrack(g, c, trackBounds)
-                if (ShowDensity.currentValue) {
-                    g as Graphics2D
-                    g.color = UIManager.getColor("Actions.Red")
-
-                    val old = g.transform
-                    g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
-                    g.transform(
-                        AffineTransform.getScaleInstance(
-                            trackBounds.width / rangex.toDouble(),
-                            trackBounds.height / density.size.toDouble(),
-                        ),
-                    )
-                    density.values.forEachIndexed { index, count ->
-                        g.drawLine(
-                            trackBounds.x,
-                            trackBounds.y + index,
-                            trackBounds.x + count,
-                            trackBounds.y + index,
-                        )
-                    }
-                    g.transform = old
-                }
-            }
-        }
-
-        init {
-            preferredSize = Dimension(30, 100)
-            setUI(customUI)
-        }
-
-        override fun updateUI() {
-            setUI(customUI)
-        }
-    }
-
-    private class Header(private val totalRows: Int) : JPanel(MigLayout("ins 0, fill, hidemode 3")) {
+    private class Header(private val totalRows: Int) : JPanel(MigLayout("ins 2, fill, hidemode 3")) {
         private val events = JLabel("Showing $totalRows of $totalRows events")
 
         val search = JXSearchField("Search")
@@ -430,25 +354,5 @@ class LogPanel(
 
     companion object {
         private val BACKGROUND = CoroutineScope(Dispatchers.Default)
-
-        private val DURATIONS = listOf(
-            Duration.ofMillis(100),
-            Duration.ofMillis(500),
-            Duration.ofSeconds(1),
-            Duration.ofSeconds(5),
-            Duration.ofSeconds(10),
-            Duration.ofSeconds(30),
-            Duration.ofMinutes(1),
-            Duration.ofMinutes(2),
-            Duration.ofMinutes(5),
-            Duration.ofMinutes(10),
-            Duration.ofMinutes(15),
-            Duration.ofMinutes(30),
-            Duration.ofHours(1),
-            Duration.ofHours(2),
-            Duration.ofHours(6),
-            Duration.ofHours(12),
-            Duration.ofDays(1),
-        )
     }
 }
