+++ src/main/kotlin/io/github/inductiveautomation/kindling/utils/FilterSidebar.kt
@@ -15,7 +15,7 @@ class FilterSidebar<T>(
         tabLayoutPolicy = SCROLL_TAB_LAYOUT
         tabsPopupPolicy = TabsPopupPolicy.asNeeded
         scrollButtonsPolicy = ScrollButtonsPolicy.never
-        tabWidthMode = TabWidthMode.equal
+        tabWidthMode = TabWidthMode.compact
         tabType = TabType.underlined
         tabHeight = 16
 
@@ -35,11 +35,16 @@ class FilterSidebar<T>(
             if (tabIndex == -1) return@attachPopupMenu null
 
             JPopupMenu().apply {
+                val filterPanel = filterPanels[tabIndex]
+
                 add(
                     Action("Reset") {
-                        filterPanels[tabIndex].reset()
+                        filterPanel.reset()
                     },
                 )
+                if (filterPanel is PopupMenuCustomizer) {
+                    filterPanel.customizePopupMenu(this)
+                }
             }
         }
